backend/foodgram/users/serializers.py
from djoser.serializers import UserSerializer
from .models import CustomUser


class CustomUserSerializer(UserSerializer):
    class Meta:
        model = CustomUser
        fields = ('email', 'id', 'username', 'first_name', 'last_name', 'is_subscribed')
        read_only_fields = 'id', 'is_subscribed'
